Update coveralls-lcov environment.
