+++ test/test_geometry.cc
@@ -5,11 +5,8 @@
 #include "../include/spica.h"
 using namespace spica;
 
-inline void EXPECT_EQ_VEC(const Vector3& v1, const Vector3& v2) {
-    EXPECT_EQ(v1.x(), v2.x());
-    EXPECT_EQ(v1.y(), v2.y());
-    EXPECT_EQ(v1.z(), v2.z());
-}
+#include "test_macros.h"
+
 
 // ------------------------------
 // Plane class test
@@ -287,43 +284,3 @@ TEST(BBoxTest, IntersectionTest) {
     EXPECT_EQ(1.0, tMin);
     EXPECT_EQ(2.0, tMax);
 }
-
-// ------------------------------
-// Trimesh class test
-// ------------------------------
-TEST(TrimeshTest, BoxIntersection) {
-    Trimesh trimesh;
-    trimesh.load(DATA_DIR + "box.ply");
-    trimesh.buildKdTreeAccel();
-
-    Ray ray(Vector3(0.0, 0.0, 100.0), Vector3(0.0, 0.0, -1.0));
-    Hitpoint hitpoint;
-    EXPECT_TRUE(trimesh.intersect(ray, &hitpoint));
-    EXPECT_EQ(99.5, hitpoint.distance());
-    EXPECT_EQ_VEC(Vector3(0.0, 0.0, 0.5), hitpoint.position());
-    EXPECT_EQ_VEC(Vector3(0.0, 0.0, 1.0), hitpoint.normal());
-}
-
-TEST(TrimeshTest, BunnyIntersection) {
-    Trimesh trimesh;
-    trimesh.load(DATA_DIR + "bunny.ply");
-    trimesh.buildKdTreeAccel();
-
-    Ray ray(Vector3(0.0, 0.0, 100.0), Vector3(0.0, 0.0, -1.0));
-
-    Hitpoint hpGT;
-    for (int i = 0; i < trimesh.numFaces(); i++) {
-        Triangle tri = trimesh.getTriangle(i);
-        Hitpoint hpTemp;
-        if (tri.intersect(ray, &hpTemp)) {
-            if (hpGT.distance() > hpTemp.distance()) {
-                hpGT = hpTemp;
-            }
-        }
-    }
-
-    Hitpoint hitpoint;
-    EXPECT_TRUE(trimesh.intersect(ray, &hitpoint));
-    EXPECT_EQ(hpGT.distance(), hitpoint.distance());
-}
-
